% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactflow.R
\name{background}
\alias{background}
\title{see also https://reactflow.dev/api-reference/components/background}
\usage{
background(...)
}
\description{
see also https://reactflow.dev/api-reference/components/background
}
