% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactflow.R
\name{controls}
\alias{controls}
\title{see also https://reactflow.dev/api-reference/components/controls}
\usage{
controls(...)
}
\description{
see also https://reactflow.dev/api-reference/components/controls
}
