% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactflow.R
\name{mini_map}
\alias{mini_map}
\title{See also https://reactflow.dev/api-reference/components/minimap}
\usage{
mini_map(...)
}
\description{
See also https://reactflow.dev/api-reference/components/minimap
}
\examples{
mini_map()

# add pan and zoom function to mini map
mini_map(pannable = NA, zoomable = NA)
}
