% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactflow.R
\name{reactflow}
\alias{reactflow}
\title{A}
\usage{
reactflow(
  nodes,
  edges,
  controls = NULL,
  mini_map = NULL,
  background = NULL,
  allow_edge_connection = TRUE,
  use_dagre = FALSE,
  dagre_direction = c("LR", "TB"),
  dagre_config = list(nodeWidth = 200, nodeHeight = 40),
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{...}{elements passed to ReactFlow (see https://reactflow.dev/api-reference/react-flow)}

\item{width}{a}

\item{height}{a}

\item{elementId}{a}

\item{message}{a}
}
\description{
A
}
